Fix typecheck and rename coreIPC for clarity
