+++ desktop/src/main/base/devtools.ts
@@ -14,47 +14,43 @@ import { getLogger } from "./logging";
 const logger = getLogger("devtools/ipc");
 
 export const devtoolsRouter = r({
-  devtools: r({
-    getSettings: proc
-      .output(devToolsConfigSchema)
-      .query(() => getDevToolsConfig()),
-    setSettings: proc
-      .input(devToolsConfigSchema)
-      .mutation(async ({ input }) => {
-        logger.info("Dev Tools settings change: " + JSON.stringify(input));
-        await saveDevToolsConfig(input);
-        IPCEvents.send("devToolsSettingsChange");
-      }),
-    throwException: proc.mutation(async () => {
-      if (!(await getDevToolsConfig()).enabled) {
-        throw new TRPCError({
-          code: "PRECONDITION_FAILED",
-          message: "Dev tools not enabled",
-        });
-      }
-      process.nextTick(() => {
-        throw new Error("Test Main Process Exception");
-      });
-    }),
-    crash: proc.mutation(async () => {
-      if (!(await getDevToolsConfig()).enabled) {
-        throw new TRPCError({
-          code: "PRECONDITION_FAILED",
-          message: "Dev tools not enabled",
-        });
-      }
-      process.crash();
-    }),
-    setEnabledIntegrations: proc
-      .input(z.array(z.string()))
-      .mutation(async ({ input }) => {
-        if (!(await getDevToolsConfig()).enabled) {
-          throw new TRPCError({
-            code: "PRECONDITION_FAILED",
-            message: "Dev tools not enabled",
-          });
-        }
-        DEV_overrideSupportedIntegrations(input as Integration[]);
-      }),
+  getSettings: proc
+    .output(devToolsConfigSchema)
+    .query(() => getDevToolsConfig()),
+  setSettings: proc.input(devToolsConfigSchema).mutation(async ({ input }) => {
+    logger.info("Dev Tools settings change: " + JSON.stringify(input));
+    await saveDevToolsConfig(input);
+    IPCEvents.send("devToolsSettingsChange");
   }),
+  throwException: proc.mutation(async () => {
+    if (!(await getDevToolsConfig()).enabled) {
+      throw new TRPCError({
+        code: "PRECONDITION_FAILED",
+        message: "Dev tools not enabled",
+      });
+    }
+    process.nextTick(() => {
+      throw new Error("Test Main Process Exception");
+    });
+  }),
+  crash: proc.mutation(async () => {
+    if (!(await getDevToolsConfig()).enabled) {
+      throw new TRPCError({
+        code: "PRECONDITION_FAILED",
+        message: "Dev tools not enabled",
+      });
+    }
+    process.crash();
+  }),
+  setEnabledIntegrations: proc
+    .input(z.array(z.string()))
+    .mutation(async ({ input }) => {
+      if (!(await getDevToolsConfig()).enabled) {
+        throw new TRPCError({
+          code: "PRECONDITION_FAILED",
+          message: "Dev tools not enabled",
+        });
+      }
+      DEV_overrideSupportedIntegrations(input as Integration[]);
+    }),
 });
