+++ desktop/src/main/coreIPC.ts
