+++ desktop/src/renderer/ipc.ts
@@ -5,7 +5,7 @@ import {
 } from "@trpc/client";
 import { ipcLink } from "electron-trpc/renderer";
 import { createTRPCReact } from "@trpc/react-query";
-import type { AppRouter } from "../main/ipcApi";
+import type { AppRouter } from "../main/coreIPC";
 import { Events } from "../common/ipcEvents";
 import { QueryKey, useQueryClient } from "@tanstack/react-query";
 import { useEffect } from "react";
